Update files
nbashotclassifier.py now has support for detecting which data set is being used; as the test set does not contain the 'SHOT_MADE_FLAG' header in the .csv, it could not be indexed. 'EVENT_TYPE' header represents the same flag and it is present in all three data sets, so it is being used instead.

The test set 'solution_no_answer.csv' was also added.